add more docs and refactor package
