+++ app/src/main/java/com/fkexample/cointicker/repo/CoinRepositoryImpl.kt
@@ -16,12 +16,22 @@ import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
 import timber.log.Timber
 
+/**
+ * Implementation of the [CoinRepository] interface.
+ * @property tickerService The service for fetching coin-related data.
+ * @property coinDao The DAO for accessing and managing coin data in the local database.
+ * @property dispatcher The CoroutineDispatcher to perform operations on.
+ */
 class CoinRepositoryImpl(
     private val tickerService: TickerService,
     private val coinDao: CoinDao,
     private val dispatcher: CoroutineDispatcher = Dispatchers.IO
 ) : CoinRepository {
 
+    /**
+     * Retrieves a flow of all coins, combining data from the API and local database.
+     * @return A flow emitting a list of [CryptoEntity] objects representing the coins.
+     */
     override fun getAllCoins(): Flow<List<CryptoEntity>> {
         return flow {
             try {
@@ -44,16 +54,24 @@ class CoinRepositoryImpl(
         }.flowOn(dispatcher)
     }
 
+    /**
+     * Adds or removes a favorite coin based on the provided [favEntity].
+     * @param favEntity The [CryptoFavEntity] representing the favorite coin.
+     */
     override suspend fun addOrRemoveFavCoin(favEntity: CryptoFavEntity) {
         val dbFav = getFavById(favEntity.assetId)
         if (dbFav != null) {
-            // This is a favorite already so remove it from the favorites
+            // This is a favorite already, so remove it from the favorites
             coinDao.deleteCoinFromFav(favEntity)
         } else {
             coinDao.insertFavCoin(favEntity)
         }
     }
 
+    /**
+     * Retrieves a flow of all favorite coins from the local database.
+     * @return A flow emitting a list of [CryptoFavEntity] objects representing the favorite coins.
+     */
     override suspend fun getAllFavoriteCoins(): Flow<List<CryptoFavEntity>> {
         return flow {
             try {
@@ -65,10 +83,20 @@ class CoinRepositoryImpl(
         }.flowOn(dispatcher)
     }
 
+    /**
+     * Retrieves the favorite coin with the specified [assetId] from the local database.
+     * @param assetId The asset ID of the favorite coin.
+     * @return The corresponding [CryptoFavEntity] if found, or null otherwise.
+     */
     override suspend fun getFavById(assetId: String): CryptoFavEntity? {
         return coinDao.getFavById(assetId)
     }
 
+    /**
+     * Retrieves the details of a coin with the specified [assetId] from the API and local database.
+     * @param assetId The asset ID of the coin.
+     * @return A flow emitting a [CryptoDetails] object representing the coin details, or null if not found.
+     */
     override suspend fun getCoinDetails(assetId: String): Flow<CryptoDetails?> {
         return flow {
             try {
@@ -103,6 +131,11 @@ class CoinRepositoryImpl(
         }.flowOn(dispatcher)
     }
 
+    /**
+     * Retrieves the coin with the specified [assetId] from the local database.
+     * @param assetId The asset ID of the coin.
+     * @return The corresponding [CryptoEntity] if found, or null otherwise.
+     */
     override suspend fun getCoinById(assetId: String): CryptoEntity? {
         return coinDao.getCoinById(assetId)
     }
