+++ app/src/main/java/com/fkexample/cointicker/utils/HttpClient.kt
@@ -3,7 +3,15 @@ package com.fkexample.cointicker.utils
 import okhttp3.OkHttpClient
 import java.util.concurrent.TimeUnit
 
+/**
+ * Utility object for creating an instance of OkHttpClient for making HTTP requests.
+ */
 object HttpClient {
+
+    /**
+     * Creates and configures an OkHttpClient instance with the necessary interceptors and timeouts.
+     * @return An instance of OkHttpClient.
+     */
     fun create(): OkHttpClient {
         return OkHttpClient.Builder()
             .addInterceptor(ApiKeyHeaderInterceptor.create())
